// src/infra/router/enums/authenticated-path.enum.ts
export enum EAuthenticatedPath {
  HOME = '',
  SETTINGS = 'configuracoes',
  ROLES = 'perfis',
  USERS = 'usuarios',
  COMPANIES = 'empresas',
  HAIRCUTS = 'cortes',
  HOURS = 'horarios',
  SCHEDULINGS = 'agendamentos',
  WORKING_DAYS = 'dias-de-trabalho',
  HOLIDAYS = 'feriados',
  ASSESSMENTS = 'avaliacoes',
  HISTORIES = 'historicos',
}